Lots of small fixes
1.  Fix Bug: 62625 - exclamation points were finding results but Elasticsearch
was whining about them.  Now we check for the whining in integration tests.
2.  Fix Bug: 62626 - if a search term started with an asterisk, had some non-
asterisk characters, then contained another asterisk then the escaping that we
did for Elasticsearch wasn't working.  This just removes the first asterisk.
3.  There was an error in the store norms code that was causing it to not
work quite right - add some more parrens to fix it. This is part of the commit
because it came up when testing the rest.

Change-Id: Idc120275172b339d4c3fe55ffc72c6e91ca648c4