Revert change to exception message finding
When implemented insource we switch to another method to find the exception
message that was fancier and maybe righter in some cases.  It broken in the
case of BulkException so we really shouldn't use it - or we should use it
more carefully.

Bug: 66780

Change-Id: I96a6130948a7ccfb57f354f1771b24c7c55f4ac2
(cherry picked from commit 7678428fec67216b88d3f8b44877c8513f3ae536)