Use the title.plain field for wildcard intitle: search
This fixes at least some of the issues detailed at:
https://phabricator.wikimedia.org/T88724

The idea is that if the user searches for an intitle: value with a wildcard
(i.e. a question mark or an asterisk), we want to query against the title.plain
field rather than the title field.  Otherwise, we stick to the title field.

Change-Id: I8de0d0871b97f9d702c764a786cfb67fe45c366c
(cherry picked from commit 5f0267a2a16712ab4d42a0f8b7507a7fcc0dd2e0)