Fails gracefully when running a job for an unknown cluster
Added a check to ensure that the cluster is known.
This can be useful in case of an emergency where the job queue has been flooded
with failed jobs due to an unreliable cluster. We may be forced to remove this
cluster from the config. Cirrus jobs created against this removed cluster will
"gracefully" fails without attempting to retry.

Without this patch an unknown cluster generates a connection to localhost and
will probably fail with 3 retries.

We should probably fail earlier but I'm not sure that the job runner will be
happy to receive an exception when calling the job constructor.

Change-Id: I418473a2f81bdb012897584028d7342feed588d3
(cherry picked from commit 7e3fd9a5dec37bd05bca13b9a74933d1b994a7ba)