Report common terms usage to javascript
The eventlogging code in the frontend needs to know if the
common terms query it turned on was actually used. This is
a bit of a hack, but should be fine for a few weeks. It
should be reverted after the test is complete.

Change-Id: I3400795c095202d08c78e2dbad2abfe2441f1374
(cherry picked from commit ab19311a7fc815c5f8bc6356b0d9f78983a143e3)