Handle ElasticaWrite job failures internally
With these write jobs not only going to production, but also
to a labsearch replica we need to be able to gracefully
handle large amounts of failures. Prior to this patch if
writes were going to the labsearch cluster but it went down
we could very well end up maxing out the memory available to
redis. Of primary concern is the abandonded queue which
would try to hold onto all the failed jobs for a week.

Rather than taking unnecssary risks this moves the handling
of failures into ElasticaWrite and bypasses the standard job
queue retry methods.

Change-Id: Idd244e627162303c9f16cdf1d9c1554d31f4e3cf
(cherry picked from 3bf3eee5f59f75431b722b6f9037b707fcd8d5ee)